add full pull and config
