+++ util/common/image.go
@@ -4,3 +4,13 @@
 */
 
 package common
+
+type DeltaImageMetadata struct {
+	ManifestSize        int64
+	ConfigSize          int64
+	StarlightHeaderSize int64
+	ContentLength       int64
+
+	Digest          string
+	StarlightDigest string
+}
